fixed but probably barely working
